// js/src/maplestory/api/response/character/characterItemEquipmentBody.ts
export type CharacterItemEquipmentEtcOptionBody = {
  str: string;
  dex: string;
  int: string;
  luk: string;
  max_hp: string;
  max_mp: string;
  attack_power: string;
  magic_power: string;
  armor: string;
  speed: string;
  jump: string;
};